+++ native/wasmex/src/lib.rs
@@ -12,5 +12,6 @@ pub mod pipe;
 pub mod printable_term_type;
 pub mod store;
 pub mod wit;
+mod component_type_conversion;
 
 rustler::init!("Elixir.Wasmex.Native");
